feat: merge 2 identity services
BREAKING CHANGE: ENG-2651
