+++ indykite_sdk_node/jest.config.js
@@ -15,10 +15,10 @@ module.exports = {
   moduleDirectories: ['src', 'node_modules'],
   coverageThreshold: {
     global: {
-      branches: 95,
-      functions: 95,
-      lines: 95,
-      statements: 95,
+      branches: 90,
+      functions: 90,
+      lines: 90,
+      statements: 90,
     },
   },
   moduleFileExtensions: ['ts', 'js', 'json'],
