+++ indykite_sdk_node/src/examples/consent.ts
@@ -4,15 +4,24 @@ const consentChallengeToken = process.env.CONSENT_CHALLENGE || 'MISSING_CONSENT_
 
 IdentityClient.createInstance()
   .then(async (sdk) => {
-    const consentChallenge = await sdk.checkConsentChallenge(consentChallengeToken);
-    consentChallenge.approveScopes(['openid', 'email']);
-    console.log('Scopes:', consentChallenge.scopes);
-    console.log('Approved scopes: ', consentChallenge.getApprovedScopeNames());
-    const { authorizationEndpoint, verifier } = await sdk.createConsentVerifier(consentChallenge);
-    console.log('Result: ', {
-      authorizationEndpoint,
-      verifier,
+    const consentChallenge = await sdk.checkConsentChallenge({
+      challenge: consentChallengeToken,
     });
+    console.log(consentChallenge?.scopes);
+
+    const consentVerifier = await sdk.createConsentVerifier({
+      consentChallenge: consentChallengeToken,
+      result: {
+        oneofKind: 'approval',
+        approval: {
+          grantedAudiences: [],
+          grantScopes: ['openid', 'email'],
+          remember: true,
+          rememberFor: '0',
+        },
+      },
+    });
+    console.log(JSON.stringify(consentVerifier, null, 2));
   })
   .catch((err) => {
     console.error(err);
