+++ indykite_sdk_node/src/examples/identity_token.ts
@@ -1,6 +1,4 @@
 import { IdentityClient } from '../sdk/identity';
-import { Property } from '../sdk/model';
-
 const userToken = process.env.USER_TOKEN || 'MISSING_TOKEN';
 
 IdentityClient.createInstance()
@@ -13,7 +11,7 @@ IdentityClient.createInstance()
     }
 
     // Get Digital Twin using token
-    const dtByToken = await sdk.getDigitalTwinByToken(userToken, ['email']);
+    const dtByToken = await sdk.getDigitalTwinByToken(userToken);
     console.log('Digital Twin By Token:', JSON.stringify(dtByToken, null, 2));
 
     if (!dtByToken.digitalTwin) {
@@ -22,41 +20,11 @@ IdentityClient.createInstance()
     }
 
     const dt = dtByToken.digitalTwin;
-
-    // Examples of getting properties
-    console.log('Get email property:', dt.getProperty('email'));
-    console.log('Get email value:', dt.getPropertyValue('email'));
-    console.log('Get email value:', dt.getProperty('email')?.value);
-    console.log('Get all email properties:', dt.getProperties('email'));
+    console.log(dt.digitalTwin?.id);
 
     const randString = Math.random().toString(36).slice(2);
     const email = `test+example_${randString}@indykite.com`;
     console.log(`Generated email: ${email}`);
-
-    // Delete all email those value starts with test+example_
-    dt.getProperties('email')
-      .filter((p) => (p.value as string).startsWith('test+example_'))
-      .forEach((p) => {
-        dt.deleteProperty(p);
-      });
-
-    // Add new property or value
-    dt.addProperty(new Property('email').withValue(email));
-
-    // Change property metadata (primary)
-    const primaryEmail = dt.getProperty('email');
-    if (primaryEmail) {
-      dt.updatePropertyMetadata(primaryEmail, false);
-    }
-
-    console.log(JSON.stringify(dt.getPatchOperation(), null, 2));
-    const patchByToken = await sdk.patchPropertiesByToken(userToken, dt);
-    console.log('Patch by token response:', JSON.stringify(patchByToken, null, 2));
-
-    // const randPwd = Math.random().toString(36).slice(2);
-    // console.log(`Random generated password: ${randPwd}`);
-    // const delByToken = await sdk.deleteDigitalTwinByToken(userToken);
-    // console.log('Delete by token', JSON.stringify(delByToken, null, 2));
   })
   .catch((err) => {
     console.error(err);
