+++ indykite_sdk_node/src/index.ts
@@ -2,10 +2,8 @@ export * as grpcIdentityModel from './grpc/indykite/identity/v1beta2/model';
 export * as grpcIdentityAPI from './grpc/indykite/identity/v1beta2/identity_management_api';
 export * as grpcIdentityAttributes from './grpc/indykite/identity/v1beta2/attributes';
 export * from './sdk/authorization';
-export * from './sdk/authorization';
 export { type InputParameters } from './sdk/authorization';
-export { IdentityClient } from './sdk/identity';
-export * from './sdk/identity_v2';
+export * from './sdk/identity';
 export { ConfigClient } from './sdk/config';
 export { ConfigClientV2 } from './sdk/config_v2';
 export { IdentityKnowledgeClient } from './sdk/knowledge';
