+++ indykite_sdk_node/src/sdk/__test__/identity/client.test.ts
@@ -1,6 +1,6 @@
 import { IdentityManagementAPIClient } from '../../../grpc/indykite/identity/v1beta2/identity_management_api.grpc-client';
 import { SdkClient } from '../../client/client';
-import { IdentityClientV2 } from '../../identity_v2';
+import { IdentityClient } from '../../identity';
 import { applicationTokenMock } from '../../utils/test_utils';
 
 afterEach(() => {
@@ -9,14 +9,14 @@ afterEach(() => {
 
 describe('when a new client is created', () => {
   describe('when no error is thrown', () => {
-    let returnedValue: IdentityClientV2;
+    let returnedValue: IdentityClient;
 
     beforeEach(async () => {
       jest
         .spyOn(SdkClient, 'createIdentityInstance')
         .mockImplementation(() => Promise.resolve({ client: {} } as SdkClient));
 
-      returnedValue = await IdentityClientV2.createInstance(JSON.stringify(applicationTokenMock));
+      returnedValue = await IdentityClient.createInstance(JSON.stringify(applicationTokenMock));
     });
 
     it('creates a new instance', () => {
@@ -24,7 +24,7 @@ describe('when a new client is created', () => {
         IdentityManagementAPIClient,
         JSON.stringify(applicationTokenMock),
       );
-      expect(returnedValue).toBeInstanceOf(IdentityClientV2);
+      expect(returnedValue).toBeInstanceOf(IdentityClient);
     });
   });
 
@@ -37,7 +37,7 @@ describe('when a new client is created', () => {
         .spyOn(SdkClient, 'createIdentityInstance')
         .mockImplementation(() => Promise.reject(error));
       try {
-        await IdentityClientV2.createInstance(JSON.stringify(applicationTokenMock));
+        await IdentityClient.createInstance(JSON.stringify(applicationTokenMock));
       } catch (err) {
         caughtError = err as Error;
       }
